Coefficient Estimates across Bootstrapped Samples

Performs a simple bootstrap of a fitted DirectEffects model by re-estimating the model with bootstrap samples.

Usage

boots_g(seqg, boots = 1000)

Arguments

seqg

A fitted sequential_g estimate, computed by sequential_g.

boots

The number of bootstrap replicates. Defaults to 1000.

Value

An object of type seqgboots which is a matrix with boots rows and columns for each coefficient in the seqg model. Use summary to provide summary statistics, such as mean and quantiles.

Examples

# \donttest{
data(ploughs)
form <- women_politics ~ plow +
 agricultural_suitability + tropical_climate + large_animals + rugged |
 years_civil_conflict + years_interstate_conflict  + oil_pc +
 european_descent + communist_dummy + polity2_2000 |
 centered_ln_inc + centered_ln_incsq
s1 <- sequential_g(form, ploughs)

out.boots <- boots_g(s1)

summary(out.boots)
#> 
#> Summary of bootstrapped coefficients: 
#> 
#>                          Estimate Std. Err.  t value    2.5 %  97.5 %
#> (Intercept)              13.23802   4.66375  2.83849  4.29674 22.9691
#> plow                     -4.42914   1.83211 -2.41750 -8.05517 -0.9433
#> agricultural_suitability  4.38532   2.34331  1.87142 -0.35748  8.8871
#> tropical_climate         -2.81405   2.33820 -1.20351 -7.57820  1.9232
#> large_animals            -0.84134   4.20320 -0.20017 -9.40630  7.0397
#> rugged                   -0.34581   0.49178 -0.70319 -1.26235  0.5908
#> 
# }