Initialize an IPW CDE estimator

Initializes the specification of a CDE estimator based on an inverse probability weighting approach.

Usage

cde_ipw(hajek = TRUE, trim = c(0.01, 0.99))

Arguments

hajek

If TRUE, normalized weights will be used as in the Hajek estimator. If FALSE, traditional IPW weights will be used.

trim

A vector of length 2 indicating what quantiles of the propensity scores should be trimmed. By default this is c(0.01, 0.99) meaning that the top and bottom 1% of propensity scores are trunctated to these quantiles. If NULL, no trimming occurs.