Data on civil wars and internal conflict from 1945-1999.

A dataset to replicate the analysis in Fearon and Laitin (2003).

A dataset to replicate the analysis in Fearon and Laitin (2003).

Usage

data(civilwar)

data(civilwar)

Format

A data frame with 6610 observations and 69 variables.

A data frame with 6610 observations and 69 variables.

Source

doi:10.1017/S0003055403000534

doi:10.1017/S0003055403000534

Details

ccode. COW country id number

country. country name

cname. abbreviated country name

cmark. 1 for first in each country series

year. start year of war/conflict

wars. number wars in progress in country year

war. 1 if war ongoing in country year

warl. lagged war, w/ 0 for start of country series

onset. 1 for civil war onset

ethonset. 1 if onset = 1 & ethwar ~= 0

durest. estimated war duration

aim. 1 = rebels aim at center, 3 = aim at exit or autonomy, 2 = mixed or ambig.

casename. Id for case, usually name of rebel group(s)

ended. war ends = 1, 0 = ongoing

ethwar. 0 = not ethnic, 1 = ambig/mixed, 2 = ethnic

waryrs. war years for each onset

pop. population, in 1000s

lpop. log of pop

polity2. revised polity score

gdpen. gdp/pop based on pwt5.6, wdi2001,cow energy data

gdptype. source/type of gdp/pop estimate

gdpenl. lagged gdpenl, except for first in country series

lgdpenl1. log of lagged gdpen

lpopl1. log population, lagged except for first in country series

region. country's region, based on MAR project

western. Dummy for Western Democracies & Japan

eeurop. Dummy for Eastern Europe

lamerica. Dummy for Latin America

ssafrica. Dummy for Sub-Saharan Africa

asia. Dummy for Asia (not including Japan)

nafrme. Dummy for North Africa/Middle East

colbrit. Former British colony

colfra. former French colony

mtnest. Estimated percent mountainous terrain

lmtnest. log of mtnest

elevdiff. high - low elevation, in meters

Oil. more than 1/3 export revenues from fuels

ncontig. noncontiguous state

ethfrac. ethnic frac. based on Soviet Atlas, plus estimates for missing in 1964

ef. ethnic fractionalization based on Fearon 2002 APSA paper

plural. share of largest ethnic group (Fearon 2002 APSA)

second. share of 2nd largest ethnic group (Fearon 2002 APSA)

numlang. number languages in Ethnologue > min(1

relfrac. religious fractionalization

plurrel. size of largest confession

minrelpc. size of second largest confession

muslim. percent muslim

nwstate. 1 in 1st 2 years of state's existence

polity2l. lagged polity2, except 1st in country series

instab. > 2 change in Polity measure in last 3 yrs

anocl. lagged anocracy (-6 < polity2l < 6)

deml. lagged democracy (polity2l > 5)

empethfrac. ethfrac coded for colonial empires

empwarl. warl coded for data with empires

emponset. onset coded for data with empires

empgdpenl. gdpenl coded for empires data

emplpopl. lpopl coded for empires data

emplmtnest. lmtnest coded for empires data

empncontig. ncontig coded for empires

empolity2l. polity2l adjusted for empires (see fn38 in paper)

sdwars. number Sambanis/Doyle civ wars in progress

sdonset. onset of Sambanis/Doyle war

colwars. number Collier/Hoeffler wars in progress

colonset. onset of Collier/Hoeffler war

cowwars. number COW civ wars in progress

cowonset. onset of COW civ war

cowwarl. 1 if COW war ongoing in last period

sdwarl. 1 if S/D war ongoing in last period

colwarl. 1 if C/H war ongoing in last period

ccode. COW country id number

country. country name

cname. abbreviated country name

cmark. 1 for first in each country series

year. start year of war/conflict

wars. number wars in progress in country year

war. 1 if war ongoing in country year

warl. lagged war, w/ 0 for start of country series

onset. 1 for civil war onset

ethonset. 1 if onset = 1 & ethwar ~= 0

durest. estimated war duration

aim. 1 = rebels aim at center, 3 = aim at exit or autonomy, 2 = mixed or ambig.

casename. Id for case, usually name of rebel group(s)

ended. war ends = 1, 0 = ongoing

ethwar. 0 = not ethnic, 1 = ambig/mixed, 2 = ethnic

waryrs. war years for each onset

pop. population, in 1000s

lpop. log of pop

polity2. revised polity score

gdpen. gdp/pop based on pwt5.6, wdi2001,cow energy data

gdptype. source/type of gdp/pop estimate

gdpenl. lagged gdpenl, except for first in country series

lgdpenl1. log of lagged gdpen

lpopl1. log population, lagged except for first in country series

region. country's region, based on MAR project

western. Dummy for Western Democracies & Japan

eeurop. Dummy for Eastern Europe

lamerica. Dummy for Latin America

ssafrica. Dummy for Sub-Saharan Africa

asia. Dummy for Asia (not including Japan)

nafrme. Dummy for North Africa/Middle East

colbrit. Former British colony

colfra. former French colony

mtnest. Estimated percent mountainous terrain

lmtnest. log of mtnest

elevdiff. high - low elevation, in meters

Oil. more than 1/3 export revenues from fuels

ncontig. noncontiguous state

ethfrac. ethnic frac. based on Soviet Atlas, plus estimates for missing in 1964

ef. ethnic fractionalization based on Fearon 2002 APSA paper

plural. share of largest ethnic group (Fearon 2002 APSA)

second. share of 2nd largest ethnic group (Fearon 2002 APSA)

numlang. number languages in Ethnologue > min(1

relfrac. religious fractionalization

plurrel. size of largest confession

minrelpc. size of second largest confession

muslim. percent muslim

nwstate. 1 in 1st 2 years of state's existence

polity2l. lagged polity2, except 1st in country series

instab. > 2 change in Polity measure in last 3 yrs

anocl. lagged anocracy (-6 < polity2l < 6)

deml. lagged democracy (polity2l > 5)

empethfrac. ethfrac coded for colonial empires

empwarl. warl coded for data with empires

emponset. onset coded for data with empires

empgdpenl. gdpenl coded for empires data

emplpopl. lpopl coded for empires data

emplmtnest. lmtnest coded for empires data

empncontig. ncontig coded for empires

empolity2l. polity2l adjusted for empires (see fn38 in paper)

sdwars. number Sambanis/Doyle civ wars in progress

sdonset. onset of Sambanis/Doyle war

colwars. number Collier/Hoeffler wars in progress

colonset. onset of Collier/Hoeffler war

cowwars. number COW civ wars in progress

cowonset. onset of COW civ war

cowwarl. 1 if COW war ongoing in last period

sdwarl. 1 if S/D war ongoing in last period

colwarl. 1 if C/H war ongoing in last period

References

Fearon, James D., and David A. Laitin (2003). Ethnicity, Insurgency, and Civil War. American Political Science Review, 97(1), 75-90. doi:10.1017/S0003055403000534

Fearon, James D., and David A. Laitin (2003). Ethnicity, Insurgency, and Civil War. American Political Science Review, 97(1), 75-90. doi:10.1017/S0003055403000534