Plot output from cdesens

Plot output from cdesens

Usage

# S3 method for class 'cdesens'
plot(
  x,
  level = 0.95,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = "Estimated ACDE",
  bty = "n",
  col = "black",
  lwd = 2,
  ci.col = "grey70",
  ref.lines = TRUE,
  ...
)

Arguments

x

output from cdesens

level

level of confidence interval to plot

xlim

the x limits (x1, x2) of the plot for the sensitivity analysis parameter, rho. Default is to use the range of rho.

ylim

the y limits of the plot for the estimated CDEs. Default is to show the all of the confidence intervals.

xlab

label for the x axis.

ylab

label for the y axis.

bty

a character string which determined the type of box which is drawn about plots. Defaults to not drawing a box. See par for more information.

col

color for the line indicating the point estimates of the bias-adjusted ACDE.

lwd

line width for the line indicating the point estimates of the bias-adjusted ACDE.

ci.col

color for the polygon that shows the confidence intervals.

ref.lines

a logical indicating whether horizontal and vertical lines at 0 should be plotted.

...

Other parameters to pass on to plot()