Data on historical plough use and the socioeconomic status of women.

A dataset to replicate the analysis in Alesina, Giuliano, and Nunn (2013).

Usage

data(ploughs)

Format

A data frame with 234 observations and 57 variables.

Source

doi: 10.1093/qje/qjt005

Details

isocode. 3-letter code for the country.

flfp2000. Female labor force participation in 2000

female_ownership. Percent of firms with female ownership (in latest survey year)

women_politics. Women in Politics in 2000, WDI

plow. Animal plow cultivation variable (v39): Using Ethnologue - pop weighted

agricultural_suitability. overall (millets, sorghum, wheat, barley, rye): share defined as suitable

tropical_climate. Frac land: tropics and subtropics: using Ethnologue - pop weighted

large_animals. presence of large animals

political_hierarchies. Jurisdictional hierarchy beyond local community (v33): Using Ethnologue - pop weighted

economic_complexity. Settlement patterns (v30)

ln_income. ln (income)

ln_income_squared. ln (income) ^2

centered_ln_inc. de-meaned ln_inc

centered_ln_incsq. de-meaned ln_inc squared

country. country name

communist_dummy. Communism indicator variable

rugged. Ruggedness (Terrain Ruggedness Index, 100 m.)

years_interstate_conflict. Years of interstate conflict, 1800-2007 - from COW

serv_va_gdp2000. Value Added in Service/GDP in 2000

polity2_2000. Polity 2 measure taken from the Polity IV dataset

oil_pc. oil production/GDP

... other variables as annotated in the source.

References

Alesina, A., Giuliano, P., & Nunn, N. (2013). On the Origins of Gender Roles: Women and the Plough. The Quarterly Journal of Economics, 128(2), 469-530. doi: 10.1093/qje/qjt005