Summarize telescope match objects

Summarize telescope match objects

Usage

# S3 method for class 'tmatch'
summary(object, ...)

Arguments

object

an object of class tmatch – results from a call to telescope_match

...

additional arguments affecting the summary produced.

Value

Returns an object of class summary.tmatch. Contains the following components

call: matched call.

m_summary: data.frame summarizes the matching ratios ({ratio}), number of units n_1, n_0, and number of matched units (matched_1, matched_0) for each treatment/mediator (term).

K: K data frame from the object telescope matching output.

L: L vector from the object telescope matching output.

a_names: character vector of the names of the treatment/mediator variables used in matching.

estimates: matrix of estimated ACDEs with and without bias correction and the estimated standard errors.

Details

summary method for tmatch objects returned by telescope_match

Returns a summary data frame containing the estimate and standard errors from the `telescope_match` object.