Data from a randomized experiment on transgender rights.

A dataset from Broockman and Kalla (2016).

Usage

data(transphobia)

Format

A data frame with 501 observations and 19 variables.

Source

doi:10.1126/science.aad9713

Details

treated. Indicator of transgender rights script (1) vs recycling script (0).

nondiscrim_law_t3. Support for transgender nondiscrimination law six weeks after treatment.

therm_trans_t2. Subjective feelings about transgender people three weeks after treatment (0 = cool, 1 = neutral, 2 = warm).

therm_obama_t1. Feeling thermometer score (0-100) for feeling warmth or coolness toward Barack Obama 3 days after treatment.

gender_norm_moral_t1. Index of moral attitudes about gender 3 days after treatment.

nondiscrim_law_t0. Baseline support for transgender nondiscrimination law.

therm_trans_t0. Baseline subjective feelings about transgender people (0 = cool, 1 = neutral, 2 = warm).

therm_obama_t0. Baseline feeling thermometer score (0-100) for feeling warmth or coolness toward Barack Obama.

gender_norm_moral_t0. Baseline index of moral attitudes about gender.

ideology_t0. Baseline measure of ideology (conservative is higher).

religious_t0. Baseline measure of religiousity.

exposure_trans_t0. Baseline indicator for personal exposure to transgender people (1) or not (0).

pid_t0. Baseline party identification (-3 = Strong Democrat, 3 = Strong Republican).

vf_democrat. Indicator for if the unit identifies as a Democrat in the voter file (1) or not (0).

vf_female. Indicator for if the unit identifies as a woman in the voter file (1) or not (0).

vf_hispanic. Indicator for if the unit identifies as Hispanic in the voter file (1) or not (0).

vf_black. Indicator for if the unit identifies as Black in the voter file (1) or not (0).

vf_age. Age of the citizen in the voter file.

nondiscrim_law_diff. Difference between nondiscrim_law_t3 and nondiscrim_law_t0

References

Broockman, D. & Kalla, J. (2016). Durably reducing transphobia: A field experiment on door-to-door canvassing. Science, 352(6282), 220-224. doi:10.1126/science.aad9713