Coefficient Estimates across Bootstrapped Samples

Performs a simple bootstrap of a fitted DirectEffects model by re-estimating the model with bootstrap samples.

Usage

boots_g(seqg, boots = 1000)

Arguments

seqg

A fitted sequential_g estimate, computed by sequential_g.

boots

The number of bootstrap replicates. Defaults to 1000.

Value

An object of type seqgboots which is a matrix with bootsrows and columns for each coefficient in the seqg model. Use summaryto provide summary statistics, such as mean and quantiles.

Examples

# \donttest{
data(ploughs)
form <- women_politics ~ plow +
 agricultural_suitability + tropical_climate + large_animals + rugged |
 years_civil_conflict + years_interstate_conflict  + oil_pc +
 european_descent + communist_dummy + polity2_2000 |
 centered_ln_inc + centered_ln_incsq
s1 <- sequential_g(form, ploughs)

out.boots <- boots_g(s1)

summary(out.boots)
#> 
#> Summary of bootstrapped coefficients: 
#> 
#>                          Estimate Std. Err.  t value    2.5 %  97.5 %
#> (Intercept)              13.00588   4.64692  2.79882  4.52723 22.2588
#> plow                     -4.42885   1.84837 -2.39608 -8.15945 -1.0985
#> agricultural_suitability  4.27893   2.48867  1.71937 -0.57298  9.0389
#> tropical_climate         -2.77231   2.41316 -1.14883 -7.69041  1.8268
#> large_animals            -0.61676   4.03505 -0.15285 -9.13419  6.1658
#> rugged                   -0.34584   0.49522 -0.69836 -1.32296  0.5941
#> 
# }