Data on health and employment outcomes measured as part of the U.S. Job Corps employment training experiment.

A dataset to replicate the analysis in Huber (2014).

A dataset to replicate the analysis in Huber (2014).

Usage

data(jobcorps)

data(jobcorps)

Format

A data frame with 10025 observations and 62 variables.

A data frame with 10025 observations and 62 variables.

Source

doi:10.1002/jae.2341

doi:10.1002/jae.2341

Details

treat. 1 = in program group. 0 = in control group.

schobef. "in school 1yr before eligibility"

trainyrbef. "training in year before Job Corps"

jobeverbef. "ever had a job before Job Corps"

jobyrbef. "job in year before job corps"

health012. "good or very good health at assignment"

health0mis. "general health at assignment missing"

pe_prb0. "physical/emotional problems at assignment"

pe_prb0mis. "missing - physical/emotional problems at assignment"

everalc. "ever abused alcohol before assignment"

alc12. "alcohol abuse one yr after assignment"

everilldrugs. "ever took illegal drugs before assignment"

age_cat. "age at application in years 16-24"

edumis. "education missing"

eduhigh. "higher education"

rwhite. "white"

everarr. "ever arrested before Job Corps"

hhsize. "household size at assignment"

hhsizemis. "household size at assignment missing"

hhinc12. "low household income at assignment"

hhinc8. "high household income at assignment"

fdstamp. "received foodstamps in yr before assignment"

welf1. "once on welfare while growing up"

welf2. "twice on welfare while growing up"

publicass. "public assistance in yr before assignment"

emplq. "worked some time 9-12 months after assignment"

emplq4full. "worked all the time in 9-12 months after assignment"

pemplq4. "proportion of weeks worked 9-12 months after assignment"

pemplq4mis. "missing - proportion of weeks worked 9-12 months after assignment"

vocq4. "in vocational training 9-12 months after assignment"

vocq4mis. "missing - in vocational training 9-12 months after assignment"

health1212. "very good or good health 1 yr after assignment"

health123. "fair health 1 yr after assignment"

pe_prb12. "1=phys/emot probs at 12 mths 0=no prob"

pe_prb12mis. "missing - physical/emotional problems 1 yr after assignment"

narry1. "number of arrests in year 1"

numkidhhf1zero. "no own kids living in household 1 yr after assignment"

numkidhhf1onetwo. "one or two own kids living in household 1 yr after assignment"

pubhse12. "1=in public housing 1 yr after assignment, 0=not in"

h_ins12a. "afdc and other transfers one yr after assignment"

h_ins12amis. "missing - afdc and other transfers one yr after assignment"

... other variables as annotated in the source.

treat. 1 = in program group. 0 = in control group.

schobef. "in school 1yr before eligibility"

trainyrbef. "training in year before Job Corps"

jobeverbef. "ever had a job before Job Corps"

jobyrbef. "job in year before job corps"

health012. "good or very good health at assignment"

health0mis. "general health at assignment missing"

pe_prb0. "physical/emotional problems at assignment"

pe_prb0mis. "missing - physical/emotional problems at assignment"

everalc. "ever abused alcohol before assignment"

alc12. "alcohol abuse one yr after assignment"

everilldrugs. "ever took illegal drugs before assignment"

age_cat. "age at application in years 16-24"

edumis. "education missing"

eduhigh. "higher education"

rwhite. "white"

everarr. "ever arrested before Job Corps"

hhsize. "household size at assignment"

hhsizemis. "household size at assignment missing"

hhinc12. "low household income at assignment"

hhinc8. "high household income at assignment"

fdstamp. "received foodstamps in yr before assignment"

welf1. "once on welfare while growing up"

welf2. "twice on welfare while growing up"

publicass. "public assistance in yr before assignment"

emplq. "worked some time 9-12 months after assignment"

emplq4full. "worked all the time in 9-12 months after assignment"

pemplq4. "proportion of weeks worked 9-12 months after assignment"

pemplq4mis. "missing - proportion of weeks worked 9-12 months after assignment"

vocq4. "in vocational training 9-12 months after assignment"

vocq4mis. "missing - in vocational training 9-12 months after assignment"

health1212. "very good or good health 1 yr after assignment"

health123. "fair health 1 yr after assignment"

pe_prb12. "1=phys/emot probs at 12 mths 0=no prob"

pe_prb12mis. "missing - physical/emotional problems 1 yr after assignment"

narry1. "number of arrests in year 1"

numkidhhf1zero. "no own kids living in household 1 yr after assignment"

numkidhhf1onetwo. "one or two own kids living in household 1 yr after assignment"

pubhse12. "1=in public housing 1 yr after assignment, 0=not in"

h_ins12a. "afdc and other transfers one yr after assignment"

h_ins12amis. "missing - afdc and other transfers one yr after assignment"

... other variables as annotated in the source.

References

Huber, M. (2014). Identifying causal mechanisms (primarily) based on inverse probability weighting. Journal of Applied Econometrics, 29(6), 920-943. doi:10.1002/jae.2341

Huber, M. (2014). Identifying causal mechanisms (primarily) based on inverse probability weighting. Journal of Applied Econometrics, 29(6), 920-943. doi:10.1002/jae.2341